==> 00-DownloadData.ipynb <==
!uv pip install git+https://github.com/eaedk/testing-zindi-package.git

# create a user object
from zindi.user import Zindian

my_username = "Kamenialexnea"
user = Zindian(username = my_username)

user.select_a_challenge(fixed_index=3)

user.which_challenge

user.download_dataset(destination="./dataset")
