==> 02-1-PlotPredictions.ipynb <==
import pandas as pd
import numpy as np
import os
import json
from PIL import Image

THR = .33
prediction = pd.read_csv('dataset/evaluations/validation.csv')
prediction.sample(5)

file_id = prediction["Image_ID"].sample(1).values[0]
file_id

FOLDER = "dataset/labels/val/"
path = os.path.join(FOLDER, file_id + ".txt")
expected = open(path, "r").readlines()
expected = [list(map(float, line.strip().split())) for line in expected]

expected

val_df = pd.read_csv("dataset/Val_df.csv")
val_df.sample(5)

raw = val_df[val_df["Image_ID"].str.contains(file_id)]

raw

predicted = prediction[prediction["Image_ID"].str.contains(file_id) & prediction["confidence"] > THR]
predicted

val_df.columns

val_df[['confidence', 'ymin', 'xmin', 'ymax', 'xmax']].describe()

from pathlib import Path

VAL_DATA_DIR = Path("dataset/images/val")
valid_files = list(VAL_DATA_DIR.glob("*.*"))
valid_files[:5]

shape = [
    Image.open(file).size for file in valid_files
]

shape[:5]

pd.DataFrame(shape, columns=["width", "height"]).describe()
