==> 05-CheckValidation.ipynb <==
import os

os.environ["CUDA_VISIBLE_DEVICES"] = "1"

# Import libraries
import pandas as pd
import os
from pathlib import Path
from tqdm import tqdm
import yaml
import matplotlib.pyplot as plt
from ultralytics import YOLO
import numpy as np
from PIL import Image
import torch

from glob import glob

# Validate the model on the validation set
BEST_PATH = sorted(glob("runs/detect/train*/weights/best.pt"))[-1]
BEST_PATH

# Load the trained YOLO model
model = YOLO(BEST_PATH)

metrics = model.val(data="data.yaml", imgsz=1024)

model("dataset/images/val/ID_a8ILTq.JPG")

model("dataset/images/val/ID_a8ILTq.JPG", imgsz=1024)

files = glob("dataset/images/val/*")
len(files)

import random

file = random.choice(files)
file

results = model(file, )

for r in results:
    print(r.boxes.xyxy, r.boxes.conf, r.boxes.cls)

from ultralytics.engine.predictor import BasePredictor

predictor = BasePredictor("runs/detect/train5/args.yaml")
predictor

predictor.args.verbose = False
predictor.args.save = False
predictor.args.save_txt = False
predictor.args.show = False

predictor.args

preds = predictor(file,)

preds

for pred in preds:
    print(pred.shape)

print([type(i) for i in preds])

for pred in preds[1]:
    print(pred.shape)

results = model(file, predictor=predictor)

for r in results:
    print(r.boxes.xyxy, r.boxes.conf, r.boxes.cls)
